requirements.txt
absl-py==1.2.0
aio-pika==8.2.4
aiofiles==22.1.0
aiogram==2.19
aiohttp==3.8.3
aiormq==6.4.2
aiosignal==1.2.0
alembic==1.8.1
amqp==5.1.1
appdirs==1.4.3
appopener==1.3
APScheduler==3.9.1
apturl==0.5.2
arabic-reshaper==2.1.3
argh==0.26.2
asgiref==3.6.0
asn1crypto==1.5.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
autogui==0.1.8
autopep8==2.0.1
awscli==1.18.105
Babel==2.9.1
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.11.1
bidict==0.22.0
billiard==3.6.4.0
blinker==1.4
blis==0.7.9
boto3==1.26.1
botocore==1.29.1
Brlapi==0.7.0
bt==0.2.9
CacheControl==0.12.11
cachetools==5.2.0
catalogue==2.0.8
celery==5.2.7
certifi==2022.5.18.1
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudpickle==2.2.0
clr-loader==0.2.4
colorama==0.4.3
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.1.0
command-not-found==0.3
confection==0.0.3
cryptography==37.0.2
cssselect2==0.6.0
cupshelpers==1.0
cycler==0.11.0
cymem==2.0.7
dask==2022.2.0
dbus-python==1.2.16
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.6
distro==1.4.0
distro-info==0.23+ubuntu1.1
dj-database-url==0.5.0
Django==3.2.5
django-admin-shortcuts==2.0.0
django-adminlte-3==0.1.6
django-allauth==0.51.0
django-baton==2.4.1
django-celery-beat==2.4.0
django-cors-headers==3.13.0
django-environ==0.8.1
django-grappelli==3.0.3
django-guardian==2.4.0
django-jazzmin==2.5.0
django-jet==1.0.8
django-jet-4==0.0.1
django-jsonfield==1.4.1
django-oauth-toolkit==2.2.0
django-rest-enumfield==0.2.0
django-rest-framework==0.1.0
django-timezone-field==5.0
djangorestframework==3.14.0
dnspython==1.16.0
docker-py==1.9.0
docopt==0.6.2
docutils==0.15.2
dominate==2.7.0
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
fbmessenger==6.0.0
ffn==0.3.6
filelock==3.8.0
fire==0.4.0
flake8==6.0.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Migrate==2.6.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
flatbuffers==22.10.26
fonttools==4.34.4
frozenlist==1.3.1
fsspec==2022.10.0
future==0.18.2
gast==0.5.3
google-api-core==2.8.2
google-api-python-client==2.56.0
google-auth==2.10.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.4
greenlet==2.0.0
grpcio==1.50.0
gunicorn==20.1.0
h11==0.9.0
h5py==3.7.0
hgicommon==1.2.0
html5lib==1.1
httplib2==0.20.4
httptools==0.1.2
humanfriendly==10.0
idna==3.3
importlib-metadata==5.0.0
importlib-resources==5.10.0
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==1.0.1
joblib==1.1.0
jsonpickle==2.2.0
jsonschema==4.16.0
jwcrypto==1.4.2
kafka-python==2.0.2
keras==2.8.0
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.4.4
kombu==5.2.4
langcodes==3.3.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==14.0.6
locket==1.0.0
lockfile==0.12.2
louis==3.12.0
lxml==4.9.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.4.1
MarkupSafe==2.0.1
matplotlib==3.5.2
mattermostwrapper==2.2
mccabe==0.7.0
monotonic==1.5
more-itertools==4.2.0
MouseInfo==0.1.3
msgpack==1.0.4
multidict==5.2.0
murmurhash==1.0.9
mysql-connector-python==8.0.31
netifaces==0.10.4
networkx==2.6.3
numpy==1.21.6
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.3.0
oscrypto==1.3.0
packaging==21.3
pamqp==3.2.1
pandas==1.4.3
pandas-datareader==0.10.0
paramiko==2.6.0
partd==1.3.0
pathtools==0.1.2
pathy==0.6.2
pdfkit==1.0.0
pexpect==4.6.0
Pillow==9.1.1
pisa==3.0.33
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
pluggy==1.0.0
preshed==3.0.8
prompt-toolkit==3.0.28
protobuf==4.21.5
psycopg==3.0.14
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycairo==1.16.2
pycodestyle==2.10.0
pycparser==2.21
pycups==1.9.73
pydantic==1.10.2
pydot==1.4.2
pyflakes==3.0.1
PyGetWindow==0.0.9
Pygments==2.3.1
PyGObject==3.36.0
pyHanko==0.13.1
pyhanko-certvalidator==0.19.5
pyinotify==0.9.6
PyJWT==2.4.0
pykwalify==1.8.0
pymacaroons==0.13.0
pymongo==3.10.1
PyMsgBox==1.0.9
PyNaCl==1.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyPdf==1.13
PyPDF2==2.2.0
PyPDF3==1.0.6
pyperclip==1.8.2
pyping==0.0.6
pypng==0.20220715.0
PyPrind==2.11.3
PyRect==0.2.0
pyRFC3339==1.1
pyrsistent==0.19.2
PyScreeze==0.1.28
python-apt==2.0.1+ubuntu0.20.4.1
python-bidi==0.4.2
python-crfsuite==0.9.8
python-crontab==2.6.0
python-dateutil==2.8.2
python-debian==0.1.36+ubuntu1.1
python-dotenv==0.20.0
python-engineio==4.3.4
python-socketio==5.7.2
python3-openid==3.2.0
python3-xlib==0.15
pythonnet==3.0.1
pytweening==1.0.4
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywhatkit==5.4
pyxdg==0.26
PyYAML==6.0
qrcode==7.3.1
questionary==1.10.0
randomname==0.1.5
rasa==3.3.0
rasa-sdk==3.3.0
redis==4.3.4
regex==2022.9.13
reportlab==3.6.10
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
requests-unixsocket==0.2.0
rocketchat-API==1.26.0
roman==2.0.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==1.0.2
scipy==1.7.3
SecretStorage==2.3.1
sentry-sdk==1.9.10
simplejson==3.16.0
six==1.16.0
sklearn-crfsuite==0.3.6
slackclient==2.9.4
smart-open==5.2.1
soupsieve==2.3.2.post1
spacy==3.4.2
spacy-legacy==3.0.10
spacy-loggers==1.0.3
SQLAlchemy==1.4.42
sqlparse==0.4.3
srsly==2.4.5
svglib==1.3.0
systemd-python==234
tabulate==0.8.10
tarsafe==0.0.3
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.3
tensorflow-addons==0.17.1
tensorflow-estimator==2.8.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.27.0
tensorflow-text==2.8.2
termcolor==2.1.0
terminaltables==3.1.10
thinc==8.1.5
threadpoolctl==3.1.0
tinycss2==1.1.1
tk==0.1.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tqdm==4.64.0
twilio==7.9.2
typeguard==2.13.3
typer==0.4.2
typing-utils==0.1.0
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
ujson==5.5.0
unattended-upgrades==0.1
uritemplate==4.1.1
uritools==4.0.0
urllib3==1.26.9
usb-creator==0.3.7
uvicorn==0.11.3
uvloop==0.17.0
vine==5.0.0
virtualenv==20.16.6
visitor==0.1.3
wadllib==1.3.3
wasabi==0.10.1
watchdog==0.8.3
wcwidth==0.2.5
webencodings==0.5.1
webexteamssdk==1.6.1
websocket-client==1.3.3
websockets==10.4
Werkzeug==1.0.1
whitenoise==5.3.0
wikipedia==1.4.0
wrapt==1.14.1
wsproto==0.15.0
xhtml2pdf==0.2.7
xhtml2pdf-legacy==3.0.38
xkit==0.0.0
yarl==1.8.1
zipp==3.10.0
